Merge from projects/mips to head by hand:
Copy the support files for the Octeon 1 CPU from sys/mips/octeon1 on
the projects/mips side to sys/mips/cavium on the head side to conform
to the other vendor code.  This code was contributed by Cavium to the
project and forward ported by Warner Losh, with some additional code
from Randal Stewart.

# I'll fix the building problems the move creates in a future commit.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@201921 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f